build.mak: default to firmware 6.0
as discussed in #89, few structs have different layout depending on the
_PSP_FW_VERSION macro.

this makes a noticable (i.e. resulting in potential memory corruption)
difference only in few structs that are supposed to be used only by
kernel experts, but it makes it clear to non-experts that the struct
layouts they program for are the latest ones.

an expert trying to target an older firmware version, can still override
the macro on the commandline.

closes #89