Add guGetDisplayState function to libgu
This function allows users to get if the display is turned on in libgu.
This can be useful for libraries like SDL, which want a way to check if
it is before trying to display the on-screen keyboard or a native
message dialog, since those do not work before libgu has been fully
initialized.