doc: search improvement and link to constants
This patch fixes the search for function like __mcount() or gprof_start().
The issue existed for functions with function attribute (__attribute__()): Doxygen didn't parsed it properly. It fixes such warnings:
warning: documented symbol '__mcount' was not declared or defined.

After changing Doxygen configuration to fix it, new warnings appears. For example:
warning: explicit link request to 'PSP_VAUDIO_FORMAT_' could not be resolved

It's fixed by using the full name of the constants. By the way, it also fixes the link to the constants in the documentation.

Initial patch based on:
https://stackoverflow.com/questions/75410662/doxygen-does-not-parse-attribute-correctly